rename shader_stage_utils.ts -> shader_builtin_utils.ts
