+++ src/webgpu/shader/validation/expression/call/builtin/textureSampleBias.spec.ts
@@ -27,7 +27,7 @@ import {
 } from '../../../../../util/conversion.js';
 import { ShaderValidationTest } from '../../../shader_validation_test.js';
 
-import { kEntryPointsToValidateFragmentOnlyBuiltins } from './shader_stage_utils.js';
+import { kEntryPointsToValidateFragmentOnlyBuiltins } from './shader_builtin_utils.js';
 
 type TextureSampleBiasArguments = {
   coordsArgType: ScalarType | VectorType;
