+++ src/webgpu/shader/validation/expression/call/builtin/textureSampleCompare.spec.ts
@@ -25,7 +25,7 @@ import {
 } from '../../../../../util/conversion.js';
 import { ShaderValidationTest } from '../../../shader_validation_test.js';
 
-import { kEntryPointsToValidateFragmentOnlyBuiltins } from './shader_stage_utils.js';
+import { kEntryPointsToValidateFragmentOnlyBuiltins } from './shader_builtin_utils.js';
 
 type TextureSampleCompareArguments = {
   coordsArgType: ScalarType | VectorType;
